# webapp/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenAPI definition",
    "description": "blackList API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/rest/users": {
      "get": {
        "tags": [
          "admin-user-resource"
        ],
        "operationId": "getAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-user-resource"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin-user-resource"
        ],
        "operationId": "deleteMany",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/rest/users/{id}": {
      "get": {
        "tags": [
          "admin-user-resource"
        ],
        "operationId": "getOne",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin-user-resource"
        ],
        "operationId": "patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonNode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin-user-resource"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDto"
                }
              }
            }
          }
        }
      }
    },
    "/rest/users/by-ids": {
      "get": {
        "tags": [
          "admin-user-resource"
        ],
        "operationId": "getMany",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rest/users/info": {
      "get": {
        "tags": [
          "admin-user-resource"
        ],
        "operationId": "info",
        "parameters": [
          {
            "name": "authentication",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Authentication"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDto"
                }
              }
            }
          }
        }
      }
    },
    "/rest/nets": {
      "get": {
        "tags": [
          "client-resource"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NetworkDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rest/nets/all": {
      "get": {
        "tags": [
          "client-resource"
        ],
        "operationId": "getAll_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NetworkDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rest/nets/updateTime": {
      "get": {
        "tags": [
          "client-resource"
        ],
        "operationId": "get_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        }
      }
    },
    "/rest/networks": {
      "get": {
        "tags": [
          "network-resource"
        ],
        "operationId": "getAll_2",
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/NetworkFilter"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelNetworkDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "network-resource"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NetworkDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "network-resource"
        ],
        "operationId": "patchMany",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonNode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "network-resource"
        ],
        "operationId": "deleteMany_1",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/rest/networks/{id}": {
      "get": {
        "tags": [
          "network-resource"
        ],
        "operationId": "getOne_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "network-resource"
        ],
        "operationId": "patch_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonNode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "network-resource"
        ],
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkDto"
                }
              }
            }
          }
        }
      }
    },
    "/rest/networks/by-ids": {
      "get": {
        "tags": [
          "network-resource"
        ],
        "operationId": "getMany_1",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NetworkDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rest/resolve": {
      "get": {
        "tags": [
          "resolve-resource"
        ],
        "operationId": "get_2",
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rest/domains": {
      "get": {
        "tags": [
          "vpn-domain-resource"
        ],
        "operationId": "getAll_3",
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/VPNDomainFilter"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelVPNDomainDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "vpn-domain-resource"
        ],
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VPNDomainDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VPNDomainDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "vpn-domain-resource"
        ],
        "operationId": "patchMany_1",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonNode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "vpn-domain-resource"
        ],
        "operationId": "deleteMany_2",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/rest/domains/{id}": {
      "get": {
        "tags": [
          "vpn-domain-resource"
        ],
        "operationId": "getOne_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VPNDomainDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "vpn-domain-resource"
        ],
        "operationId": "patch_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonNode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VPNDomainDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "vpn-domain-resource"
        ],
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VPNDomainDto"
                }
              }
            }
          }
        }
      }
    },
    "/rest/domains/by-ids": {
      "get": {
        "tags": [
          "vpn-domain-resource"
        ],
        "operationId": "getMany_2",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VPNDomainDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdminUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "JsonNode": {
        "type": "object",
        "properties": {
          "integralNumber": {
            "type": "boolean"
          },
          "double": {
            "type": "boolean"
          },
          "valueNode": {
            "type": "boolean"
          },
          "floatingPointNumber": {
            "type": "boolean"
          },
          "bigInteger": {
            "type": "boolean"
          },
          "nodeType": {
            "type": "string",
            "enum": [
              "ARRAY",
              "BINARY",
              "BOOLEAN",
              "MISSING",
              "NULL",
              "NUMBER",
              "OBJECT",
              "POJO",
              "STRING"
            ]
          },
          "float": {
            "type": "boolean"
          },
          "int": {
            "type": "boolean"
          },
          "long": {
            "type": "boolean"
          },
          "textual": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          },
          "missingNode": {
            "type": "boolean"
          },
          "pojo": {
            "type": "boolean"
          },
          "number": {
            "type": "boolean"
          },
          "boolean": {
            "type": "boolean"
          },
          "null": {
            "type": "boolean"
          },
          "array": {
            "type": "boolean"
          },
          "binary": {
            "type": "boolean"
          },
          "containerNode": {
            "type": "boolean"
          },
          "short": {
            "type": "boolean"
          },
          "bigDecimal": {
            "type": "boolean"
          },
          "object": {
            "type": "boolean"
          }
        }
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
      "Authentication": {
        "type": "object",
        "properties": {
          "principal": {
            "$ref": "#/components/schemas/Object"
          },
          "authenticated": {
            "type": "boolean"
          },
          "credentials": {
            "$ref": "#/components/schemas/Object"
          },
          "name": {
            "type": "string"
          },
          "details": {
            "$ref": "#/components/schemas/Object"
          },
          "authorities": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "NetworkDto": {
        "type": "object",
        "properties": {
          "network": {
            "type": "string"
          },
          "manual": {
            "type": "boolean"
          },
          "active": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "imported": {
            "type": "boolean"
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          },
          "comment": {
            "type": "string"
          }
        }
      },
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PagedModelNetworkDto": {
        "type": "object",
        "properties": {
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NetworkDto"
            }
          }
        }
      },
      "NetworkFilter": {
        "type": "object",
        "properties": {
          "networkContains": {
            "type": "string"
          },
          "manual": {
            "type": "boolean"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "VPNDomainDto": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PagedModelVPNDomainDto": {
        "type": "object",
        "properties": {
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VPNDomainDto"
            }
          }
        }
      },
      "VPNDomainFilter": {
        "type": "object",
        "properties": {
          "domainContains": {
            "type": "string"
          },
          "commentContains": {
            "type": "string"
          }
        }
      }
    }
  }
}